+++ exch/emsmdb/rop_ext.cpp
@@ -1997,7 +1997,7 @@ pack_result rop_ext_make_rpc_ext(const void *pbuff_in, uint32_t in_len,
 			rpc_header_ext.flags &= ~RHE_FLAG_COMPRESSED;
 		} else {
 			auto compressed_len = lzxpress_compress(ext_buff.get(), subext.m_offset, tmp_buff.get(), ext_buff_size);
-			if (compressed_len <= 0 || compressed_len >= subext.m_offset) {
+			if (compressed_len <= 0 || static_cast<size_t>(compressed_len) >= subext.m_offset) {
 				/* if we can not get benefit from the
 					compression, unmask the compress bit */
 				rpc_header_ext.flags &= ~RHE_FLAG_COMPRESSED;
